add new plugin to load external gain solution
